Split init checker in two: one sub-, one super-checker; init checker now no longer needs to cache errors for parent to report them
