+++ checker/src/main/java/org/checkerframework/checker/initialization/InitializationFieldAccessChecker.java
@@ -1,0 +1,8 @@
+package org.checkerframework.checker.initialization;
+
+import org.checkerframework.common.basetype.BaseTypeChecker;
+
+public class InitializationFieldAccessChecker extends BaseTypeChecker {
+
+    public InitializationFieldAccessChecker() {}
+}
