+++ checker/src/main/java/org/checkerframework/checker/initialization/InitializationFieldAccessVisitor.java
@@ -1,13 +1,14 @@
-package org.checkerframework.checker.nullness;
+package org.checkerframework.checker.initialization;
 
 import com.sun.source.util.TreePath;
 
 import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.common.basetype.BaseTypeVisitor;
 
-public class NullnessVisitor extends BaseTypeVisitor<NullnessAnnotatedTypeFactory> {
+public class InitializationFieldAccessVisitor
+        extends BaseTypeVisitor<InitializationFieldAccessAnnotatedTypeFactory> {
 
-    public NullnessVisitor(BaseTypeChecker checker) {
+    public InitializationFieldAccessVisitor(BaseTypeChecker checker) {
         super(checker);
     }
