+++ checker/src/main/java/org/checkerframework/checker/nullness/NonNullValue.java
@@ -19,7 +19,7 @@ import javax.lang.model.type.TypeMirror;
  * Behaves just like {@link CFValue}, but additionally tracks whether at this point {@link PolyNull}
  * is known to be {@link NonNull} or {@link Nullable} (or not known to be either)
  */
-public class NullnessValue extends CFAbstractValue<NullnessValue> {
+public class NonNullValue extends CFAbstractValue<NonNullValue> {
 
     /** True if, at this point, {@link PolyNull} is known to be {@link NonNull}. */
     protected boolean isPolyNullNonNull;
@@ -34,16 +34,16 @@ public class NullnessValue extends CFAbstractValue<NullnessValue> {
      * @param annotations the annotations
      * @param underlyingType the underlying type
      */
-    public NullnessValue(
-            CFAbstractAnalysis<NullnessValue, ?, ?> analysis,
+    public NonNullValue(
+            CFAbstractAnalysis<NonNullValue, ?, ?> analysis,
             AnnotationMirrorSet annotations,
             TypeMirror underlyingType) {
         super(analysis, annotations, underlyingType);
     }
 
     @Override
-    public NullnessValue leastUpperBound(NullnessValue other) {
-        NullnessValue result = super.leastUpperBound(other);
+    public NonNullValue leastUpperBound(NonNullValue other) {
+        NonNullValue result = super.leastUpperBound(other);
 
         AnnotationMirror resultNullableAnno =
                 analysis.getTypeFactory().getAnnotationByClass(result.annotations, Nullable.class);
@@ -59,7 +59,7 @@ public class NullnessValue extends CFAbstractValue<NullnessValue> {
                             && this.containsNullableOrPolyNull())) {
                 result.annotations.remove(resultNullableAnno);
                 result.annotations.add(
-                        ((NullnessAnnotatedTypeFactory) analysis.getTypeFactory()).POLYNULL);
+                        ((NonNullAnnotatedTypeFactory) analysis.getTypeFactory()).POLYNULL);
             }
         }
         return result;
